add API category

+++ app/Http/Controllers/ProductController.php
@@ -2,11 +2,11 @@
 
 namespace App\Http\Controllers;
 
+use App\Http\Resources\CategoryResource;
 use Illuminate\Http\Request;
 use App\Models\Product;
 use App\Http\Resources\ProductResource;
-use GuzzleHttp\Handler\Proxy;
-use Illuminate\Support\Facades\DB;
+use App\Models\User;
 use Illuminate\Support\Facades\Storage;
 use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Facades\Auth;
@@ -21,11 +21,11 @@ class ProductController extends Controller
     {
         //
         $products = Product::paginate(15);
-        
+
         return ProductResource::collection($products);
    // $pr=  DB::select('select * from products where userID =?',[73]);
 
-       
+
     }
 
     /**
@@ -46,17 +46,22 @@ class ProductController extends Controller
      */
     public function store(Request $request)
     {
-        //
-        // Auth::check() ? Auth::user()->id : null;
-        // $id = Auth::id(); 
+        if(Auth::check())
+        {
+            $id = Auth::id();
+        }
+        else{
+            return response()->json(['status'=>"Login to Continue"]);
+        }
+
+      
         $rule= array(
             'pname'=>'required',
              'description'=>'required',
              'quantity'=>'required|integer|min:1',
              'price'=>'required|integer|min:0',
              'photoURL'=>'required|image|max:2000',
-             'cate_id'=>'required|integer',    
-             'userID' =>'required|integer'
+             'cate_id'=>'required|integer',
     );
     $validator =  Validator::make($request->all(),$rule);
     if($validator->fails())
@@ -72,7 +77,8 @@ class ProductController extends Controller
         'quantity'=>$request->quantity,
         'price'=>$request->price,
         'photoURL'=>$path,
-        'userID'=>$request->userID,
+        // 'userID'=>$request->userID,
+        'userID'=>$id,
         'cate_id'=>$request->cate_id
     ];
     $newProduct = Product::create($dataInsert);
@@ -91,8 +97,6 @@ class ProductController extends Controller
     {
         $product = Product::findOrFail($id);
         return new ProductResource($product);
-       // $product = Product::where('id', '=', $id)->get();
-       // return ProductResources::collection($product);
     }
 
     /**
@@ -115,8 +119,15 @@ class ProductController extends Controller
      */
     public function update(Request $request, $id)
     {
-        Auth::check() ? Auth::user()->id : null;
-        $id = Auth::id();
+        $productUpdate = Product::firstOrFail($id);
+       // dd($request);
+       if(Auth::check())
+        {
+            $userid = Auth::id();
+        }
+        else{
+            return response()->json(['status'=>"Login to Continue"]);
+        }
         $rule= array(
             'pname'=>'required',
              'description'=>'required',
@@ -125,14 +136,13 @@ class ProductController extends Controller
              'photoURL'=>'required|image|max:2000',
              'cate_id'=>'required|integer',
              'userID'=>'required|integer|min:0'
-        
+
     );
     $validator =  Validator::make($request->all(),$rule);
     if($validator->fails())
     {
             return $validator->errors();
     }
-    $productUpdate = Product::find($id);
     $path = Storage::disk('s3')->put('images/originals', $request->photoURL, 'public');
     $dataInsert = [
         'pname'=>$request->pname,
@@ -140,7 +150,7 @@ class ProductController extends Controller
         'quantity'=>$request->quantity,
         'price'=>$request->price,
         'photoURL'=>$path,
-        'userID'=>$request->userID,
+        'userID'=> $userid,
         'cate_id'=>$request->cate_id
     ];
     $productUpdate->update($dataInsert);
@@ -155,8 +165,55 @@ class ProductController extends Controller
      */
     public function destroy($id)
     {
-        $productDelete = Product::find($id);
+
+        if(Auth::check())
+        {
+            $userid = Auth::id();
+        }
+        else{
+            return response()->json(['status'=>"Login to Continue"]);
+        }
+        $productDelete = Product::firstOrFail($id);
+        if($productDelete->user->id != $userid ) return response()->json(['error' => 'Unauthorized'], 401);
+
         $productDelete->delete();
         return $id;
     }
+    public function suggestProdByCate(Request $request)
+    {
+        
+        $rule= array(
+            'id'=>'required'
+        );
+
+       $validator =  Validator::make($request->all(),$rule);
+
+        if($validator->fails())
+        {
+              //  return $validator->errors();
+              return $request;
+        
+        }
+        $product = Product::where('id',$request->id)->first();
+        $productSugg = Product::where([['cate_id','=',$product->cate_id],['id','!=',$product->id]])->paginate(10);
+        return $productSugg;
+
+    }
+    public function suggestProdByUser(Request $request)
+    {
+        $rule= array(
+            'id'=>'required'
+        );
+
+       $validator =  Validator::make($request->all(),$rule);
+
+        if($validator->fails())
+        {
+                return $validator->errors();
+        }
+        $product = Product::where('id',$request->id)->first();
+        $productSugg = Product::where([['userID','=',$product->userID],['id','!=',$product->id]])->paginate(10);
+        return $productSugg;
+    }
+
 }
